+++ src/App/App.css
@@ -1,55 +1,17 @@
 .App {
   text-align: center;
-  * { margin: 0; padding: 0; }
-           p { padding: 10px; }
 }
 
-#left { 
+.left { 
   position: absolute; 
   left: 0; 
   top: 0; 
   width: 50%; 
 }
 
-#right { 
+.right { 
   position: absolute; 
   right: 0; 
   top: 0; 
   width: 50%; 
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.App-header {
-  background-color: #fbfcff;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: black;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.text {
-  margin-left: auto;
-  margin-right: auto;
-  display: block;
 }
