+++ src/App/App.tsx
@@ -1,34 +1,15 @@
 import React, { useState } from 'react';
-//import logo from './logo.svg';
 import './App.css';
-import {PrintText} from '../ReturnText/ReturnText';
-import {MakeText} from '../MakeTheText/MakeTheText';
+import {PrintText} from '../PrintText';
+import {MakeText} from '../MakeText';
 
 const App: React.FC = () => {
   const [text, setText] = useState();
+
   return (
     <div className="App">
-      <header className="App-header">
-        {/*<img src={logo} className="App-logo" alt="ya tyt" />*/}
-        <div id="left">
-          <PrintText text={text}/>
-        </div>
-        {/*<p>
-          Edit <code>src/App/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>*/}
-        <div id="right">
-          <p><b>Введите текст:</b></p>
-          <MakeText onSubmit = {setText}/>
-        </div>
-      </header>
+          <PrintText className="left" text={text}/>
+          <MakeText className="right" onSubmit = {setText}/>
     </div>
   );
 }
