+++ src/MakeText/index.ts
@@ -1,0 +1,3 @@
+import {MakeText as Component} from './MakeText'
+
+export const MakeText = Component;
