--- src/MakeTheText/MakeTheText.css
@@ -1,6 +1,0 @@
-.button {
-}
-
-.text {
-	display: block;
-}
