+++ src/PrintText/index.ts
@@ -1,0 +1,3 @@
+import {PrintText as Component} from './PrintText'
+
+export const PrintText = Component;
