--- src/ReturnText/ReturnText.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export function PrintText(props: any){
-    return <h1>{ props.text }</h1>;
-}
